Renamed and reduced end to end test
